I got in over my head with this refactor. I will ask Chef for help on my own time in order to progress toward something that does not break all my code.
