--- app/controllers/media_controller.rb
@@ -1,2 +1,0 @@
-class MediaController < ApplicationController
-end
