--- app/helpers/media_helper.rb
@@ -1,2 +1,0 @@
-module MediaHelper
-end
