+++ app/models/medium.rb
@@ -6,4 +6,18 @@ class Medium < ActiveRecord::Base
   def add_a_vote
     self.rank += 1
   end
+
+  def find_all_media
+    @all_media = self.all.order(rank :desc)
+  end
+
+  def find_media
+    # find a specific instance based on params
+    @media = self.find(params[:id])
+  end
+
+  def create_medium
+    # medium is the new instance
+    @medium = self.new
+  end
 end
