--- spec/controllers/media_controller_spec.rb
@@ -1,5 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe MediaController, type: :controller do
-
-end
